fix(rust): credential example successfully verifies
