+++ implementations/rust/examples/credentials/Cargo.toml
@@ -19,6 +19,7 @@ name = "holder"
 path = "src/verifier.rs"
 name = "verifier"
 
+
 [dependencies]
 hex = "0.4"
 ockam = { version = "*", path = "../../ockam/ockam"}
